+++ src/app/[locale]/(main)/influencer/[influencerId]/review/[reviewId]/_components/ReviewCommentCard.tsx
@@ -1,14 +1,10 @@
 import { cn } from '@/lib/utils';
 
 import { useTranslations } from 'next-intl';
-import { useModalStore } from '@/stores/modalStore';
-import { useInformationToast } from '@/hooks/useInformationToast';
+import { useMyCommentMutations } from '../_hooks/useMyCommentMutations';
 
 import { VscChromeClose } from 'react-icons/vsc';
-import MessageBox from '@/components/common/MessageBox';
-
 import { formatDateToYYMMDD, parseISOToDate } from '@/lib/date';
-import { useDeleteInfluencerReviewComment } from '@/hooks/queries/useReviewService';
 
 interface DeleteCommentProps {
   isMyComment: boolean;
@@ -45,41 +41,7 @@ const ReviewCommentCard = ({
   isDeleted,
   isMyComment,
 }: ReviewCommentCardProps) => {
-  const t = useTranslations('review_page');
-
-  // 댓글 삭제 로직 훅으로 빼기, 삭제 시 리액트쿼리 캐시데이터 수정하기
-  const openModal = useModalStore((state) => state.openModal);
-  const { showConfirmToast, showErrorToast } = useInformationToast();
-
-  const deleteCommentMutation = useDeleteInfluencerReviewComment();
-  const handleDeleteComment = async (influencerId: number, reviewId: number, commentId: number) => {
-    openModal(
-      <MessageBox
-        title={t('정말 삭제하시겠어요?')}
-        description={t('삭제한 한줄리뷰는 다시 복구할 수 없어요')}
-        buttons={[
-          {
-            text: t('삭제'),
-            color: 'orange',
-            onClick: async () => {
-              try {
-                await deleteCommentMutation.mutateAsync({
-                  influencerId,
-                  reviewId,
-                  commentId,
-                });
-                showConfirmToast('댓글이 정상적으로 삭제되었어요.');
-              } catch {
-                showErrorToast('댓글 삭제에 실패했어요', '다시 시도해 주세요');
-              }
-            },
-          },
-          { text: t('취소'), color: 'gray' },
-        ]}
-      />,
-    );
-  };
-
+  const { handleDeleteComment } = useMyCommentMutations();
   return (
     <div className="flex flex-col justify-center gap-y-2.5 pb-[15px]">
       {isDeleted ? (
