+++ src/components/domain/influencer/MetricsText.tsx
@@ -1,5 +1,5 @@
-import { Separator } from '@/components/ui/separator';
 import { useTranslations } from 'next-intl';
+import { Separator } from '@/components/ui/separator';
 
 interface MetricsTextProps {
   contentsRating: number;
